# senbara-rest/api/openapi/v1/openapi.yaml
openapi: 3.1.1
info:
  title: Senbara REST API v1
  summary: Personal ERP REST API using the Go stdlib, OIDC and PostgreSQL
  description: REST API for a simple personal ERP web application built with the Go standard library, OpenID Connect authentication and PostgreSQL data storage. Designed as a reference for modern REST API development with Go.
  termsOfService: https://felicitas.pojtinger.com/tos/dynamic # Filled out at runtime
  x-privacy-policy: https://felicitas.pojtinger.com/privacy/dynamic # Filled out at runtime
  contact:
    name: Felicitas Pojtinger # Filled out at runtime
    url: https://felicitas.pojtinger.com/imprint # Filled out at runtime
    email: felicitas@pojtinger.com # Filled out at runtime
  license:
    name: AGPL-3.0
    url: https://github.com/pojntfx/senbara/blob/main/LICENSE
  version: 0.1.0
servers:
  - url: http://localhost:1337/ # Filled out at runtime
    description: Local development server # Filled out at runtime
...
paths:
...
  /journal:
    get:
      tags:
        - journal
      summary: List all journal entries
      operationId: getJournalEntries
      security:
        - oidc: []
      responses:
        "200":
          description: Journal entries retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
        "403":
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
        "500":
          description: An internal error occurred
          content:
            text/plain:
              schema:
                type: string
    post:
      tags:
        - journal
      summary: Create a new journal entry
      operationId: createJournalEntry
      security:
        - oidc: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                body:
                  type: string
                rating:
                  type: integer
                  format: int32
              required:
                - title
                - body
                - rating
      responses:
        "200":
          description: Journal entry created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "403":
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
        "500":
          description: An internal error occurred
          content:
            text/plain:
              schema:
                type: string
  /journal/{id}:
    get:
      tags:
        - journal
      summary: Get a specific journal entry
      operationId: getJournalEntry
      security:
        - oidc: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Journal entry retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "403":
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
        "500":
          description: An internal error occurred
          content:
            text/plain:
              schema:
                type: string
    delete:
      tags:
        - journal
      summary: Delete a journal entry
      operationId: deleteJournalEntry
      security:
        - oidc: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Journal entry deleted successfully
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "403":
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
        "500":
          description: An internal error occurred
          content:
            text/plain:
              schema:
                type: string
    put:
      tags:
        - journal
      summary: Update a journal entry
      operationId: updateJournalEntry
      security:
        - oidc: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                body:
                  type: string
                rating:
                  type: integer
                  format: int32
              required:
                - title
                - body
                - rating
      responses:
        "200":
          description: Journal entry updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "403":
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
        "500":
          description: An internal error occurred
          content:
            text/plain:
              schema:
                type: string
...
components:
  schemas:
...
    JournalEntry:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        date:
          type: string
          format: date-time
        body:
          type: string
        rating:
          type: integer
          format: int32
...
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: /.well-known/openid-configuration # Filled out at runtime
      x-oidc-dcr-initial-access-token-portal-url: ~ # Filled out at runtime